# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Expense Tracker Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseListResponse"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /income:
    get:
      operationId: getIncomeList
      tags: [income]
      summary: List income entries
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of income
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeListResponse"
    post:
      operationId: createIncome
      tags: [income]
      summary: Create an income entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
  /income/{id}:
    put:
      operationId: updateIncome
      tags: [income]
      summary: Update an income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
    delete:
      operationId: deleteIncome
      tags: [income]
      summary: Delete an income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List financial goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a financial goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a financial goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        category:
          type: string
        note:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
      required: [id, userId, amount, category, date, createdAt]
    ExpenseInput:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
        note:
          type: string
        date:
          type: string
      required: [amount, category, date]
    ExpenseUpdate:
      type: object
      properties:
        amount:
          type: number
        category:
          type: string
        note:
          type: string
        date:
          type: string
    ExpenseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    Income:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        category:
          type: string
        source:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
      required: [id, userId, amount, category, date, createdAt]
    IncomeInput:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
        source:
          type: string
        note:
          type: string
        date:
          type: string
      required: [amount, category, date]
    IncomeUpdate:
      type: object
      properties:
        amount:
          type: number
        category:
          type: string
        source:
          type: string
        note:
          type: string
        date:
          type: string
    IncomeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Income"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, title, targetAmount, currentAmount, createdAt]
    GoalInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        targetAmount:
          type: number
          minimum: 0
        currentAmount:
          type: number
        deadline:
          type: string
      required: [title, targetAmount]
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
...