# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  userId: number;
  amount: number;
  category: string;
  /** @nullable */
  note?: string | null;
  date: string;
  createdAt: string;
}