# lib/api-zod/src/generated/types/expenseListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { Expense } from './expense';

export interface ExpenseListResponse {
  data: Expense[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}