# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  amount?: number;
  category?: string;
  note?: string;
  date?: string;
}